Use custom assert macro & add network description system
Allow us to create a sequential description which we can then use to construct a sequential network. We use some epic std::variant magic here to make this system easy to use. Eventually we will describe a network with some kind of external json or xml file.

Additionally, we also allow for no activation for fully connected layers.

The regression example has no been improved to model a generic lambda function which is specified at the top of its scope.